Refractor Lottie animation
